Support and Resistance included and error and retry reports

+++ src/Trading_View.java
@@ -28,10 +28,13 @@ public class Trading_View {
 	public static WebDriver chrome = null;
 	private static List<NSEScript> records = null;
 	private static String lastRecordSymbol = null;
+	private static List<String> errorList = new ArrayList<String>();
+
+	private static boolean overwriteFlag = true;
 
 	// Nifty 50
 //	private static String inputFileName = "/Users/anbu/Downloads/ind_nifty50list.csv";
-//	private static String outFilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/nifty50_report.out";	
+//	private static String outFilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/nifty50_report.out";
 //	private static String retryfilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/nifty50_report_retry.out";
 
 	// Nifty 500
@@ -40,33 +43,40 @@ public class Trading_View {
 //	private static String retryfilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/nifty500_report_retry.out";
 
 	// Portfolio
-//	private static String inputFileName = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/portfolio.txt";
-//	private static String outFilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/portfolio_report.out";
-//	private static String retryfilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/portfolio_report_retry.out";
+	private static String inputFileName = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/portfolio.txt";
+	private static String outFilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/portfolio_report.out";
+	private static String retryfilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/portfolio_report_retry.out";
 
 	// Watchlist N500 Buy
-	private static String inputFileName = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist_Buy.txt";
-	private static String outFilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist_Buy.out";
-	private static String retryfilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist_Buy_retry.out";
+//	private static String inputFileName = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist_Buy.txt";
+//	private static String outFilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist_Buy.out";
+//	private static String retryfilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist_Buy_retry.out";
 
 	// Watchlist - General
 //	private static String inputFileName = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist.txt";
 //	private static String outFilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist_report.out";
 //	private static String retryfilename = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/Watchlist_report_retry.out";
 
+	
 	private static String chromeDriverLocation = "/Users/anbu/Downloads/chromedriver";
 	private static String screenshotLocation = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/sel.png";
 	private static String htmlLocation = "/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/report.html";
 	private static boolean retriedFlag = false;
 	private static List<NSEScript> retryRecords = new ArrayList<NSEScript>();
 	private static List<NSEScript> outRecords = new ArrayList<NSEScript>();
+	private static List<NSEScript> buyRecords = new ArrayList<NSEScript>();
+	private static List<NSEScript> sellRecords = new ArrayList<NSEScript>();
 
 	private static long sleepTime = 3000;
+	private static String SEP_COMMA = ",";
 	private static String buySpanStr = "//span[contains(@class, 'counterNumber-3l14ys0C-') and contains(@class, 'brandColor-1WP1oBmS-')]";
 	private static String sellSpanStr = "//span[contains(@class, 'counterNumber-3l14ys0C-') and contains(@class, 'redColor-Hpg7doOR-')]";
 	private static String snrClssNm = "itemContent-OyUxIzTS-";
 	private static String dySmCssStr = "div.summary-72Hk5lHE- span.speedometerSignal-pyzN--tL-";
 	private static String dySmXpathStr = "//*[@id=\"technicals-root\"]/div/div/div[2]/div[2]/span[2]";
+	private static String camarillaS3XpathStr="//*[@id=\"technicals-root\"]/div/div/table/tbody/tr[2]/td[4]";
+	private static String camarillaR3XpathStr="//*[@id=\"technicals-root\"]/div/div/table/tbody/tr[8]/td[4]";
+	
 
 	private static class NSEScript {
 		private String companyName;
@@ -77,6 +87,8 @@ public class Trading_View {
 		private String day;
 		private String week;
 		private String month;
+		private String resistance;
+		private String support;
 
 		public NSEScript(String companyName, String industry, String symbol, String series, String isin) {
 			this.companyName = companyName;
@@ -86,6 +98,15 @@ public class Trading_View {
 			this.isin = isin;
 		}
 
+		public String getResistance() {
+			return resistance;
+		}
+		
+		public String getSupport() {
+			return support;
+		}
+		
+		
 		public String getCompanyName() {
 			return companyName;
 		}
@@ -122,6 +143,15 @@ public class Trading_View {
 			this.day = day;
 		}
 
+		public void setResistance(String day) {
+			this.resistance = day;
+		}
+		
+		public void setSupport(String day) {
+			this.support = day;
+		}
+		
+		
 		public void setWeek(String week) {
 			this.week = week;
 		}
@@ -161,11 +191,10 @@ public class Trading_View {
 
 		chrome.get(strUrl);
 		System.out.println("Title: " + chrome.getTitle());
-		if(chrome.getTitle().contains("Page Not Found"))
-		{			
+		if (chrome.getTitle().contains("Page Not Found")) {
 			throw new Exception("Page Not Found");
 		}
-		
+
 //		// Load page & take screenshot of full-screen page
 //		File scrFile = ((TakesScreenshot) chrome).getScreenshotAs(OutputType.FILE);
 //		FileUtils.copyFile(scrFile, new File("/Users/anbu/Downloads/phantomjs-2.1.1-macosx/work/trading_view/sel.png"));
@@ -176,10 +205,16 @@ public class Trading_View {
 		// speedometerSignal-pyzN--tL- neutralColor-15OoMFX9-
 		String daySummary = chrome.findElement(By.cssSelector(dySmCssStr)).getText();
 		String daySummary1 = chrome.findElement(By.xpath(dySmXpathStr)).getText();
-
-		System.out.println("Day Summary: " + daySummary);
-		System.out.println("Day Summary1: " + daySummary1);
+		
+		//Camarilla Support & Resistance data		
+		String camS3 = chrome.findElement(By.xpath(camarillaS3XpathStr)).getText();
+		String camR3 = chrome.findElement(By.xpath(camarillaR3XpathStr)).getText();
+		
+//		System.out.println("Day Summary: " + daySummary);
+		System.out.println("Day Summary1: " + daySummary1 + "\tResistance: " + camR3 + "\tSupport: " + camS3);
 		script.setDay(daySummary1);
+		script.setResistance(camR3);
+		script.setSupport(camS3);
 
 		// Buy
 		// Spans
@@ -211,8 +246,8 @@ public class Trading_View {
 
 		if ((daySummary1 == null || daySummary1.trim().isEmpty())
 				&& (dayScenario != null && dayScenario.getText().contains("day"))) {
-			System.out.println("Day: ");
-			script.setDay(crawlNextPage(dayScenario, buyVal));
+//			System.out.println("Day: ");
+			script.setDay(crawlNextPage(dayScenario, buyVal));			
 		}
 
 		System.out.println("OUT:\t" + script.getSymbol() + "," + script.getDay() + "," + script.getWeek() + ","
@@ -224,6 +259,16 @@ public class Trading_View {
 			retriedFlag = false;
 		} else {
 			outRecords.add(script);
+		}
+		
+		//Daily and Weekly Buy calls
+		if(script.getDay().contains("BUY") && script.getWeek().contains("BUY"))
+		{
+			buyRecords.add(script);
+		}
+		else if(script.getDay().contains("SELL") && script.getWeek().contains("SELL"))
+		{
+			sellRecords.add(script);
 		}
 
 	}
@@ -259,8 +304,10 @@ public class Trading_View {
 		JavascriptExecutor jse = (JavascriptExecutor) chrome;
 		jse.executeScript("window.scrollBy(0,250)", "");
 		jse.executeScript("window.scrollBy(0,250)", "");
-		writeHtml();
-		saveScreenshot();
+		
+//		DEBUG
+//		writeHtml();
+//		saveScreenshot();
 
 		Integer buyVal = 0;
 //		List<WebElement> buySpans = chrome.findElements(By.className("span.counterNumber-3l14ys0C-.brandColor-1WP1oBmS-"));
@@ -310,10 +357,11 @@ public class Trading_View {
 
 		cmpBuyVal = buyVal;
 
-		String summary = chrome.findElement(By.cssSelector("div.summary-72Hk5lHE- span.speedometerSignal-pyzN--tL-"))
+		//"div.summary-72Hk5lHE- span.speedometerSignal-pyzN--tL-"
+		String summary = chrome.findElement(By.cssSelector(dySmCssStr))
 				.getText();
 		System.out.println(summary);
-
+		
 		return summary;
 	}
 
@@ -388,12 +436,18 @@ public class Trading_View {
 		try {
 
 			File f = new File(filename);
-			FileWriter writer = new FileWriter(f, true);
+			FileWriter writer = null;
+
+			if (overwriteFlag) {
+				writer = new FileWriter(f, false);
+			} else {
+				writer = new FileWriter(f, true);
+			}
 
 			for (Iterator iterator = records.iterator(); iterator.hasNext();) {
 				NSEScript nseScript = (NSEScript) iterator.next();
-				writer.write(nseScript.getSymbol() + "," + nseScript.getDay() + "," + nseScript.getWeek() + ","
-						+ nseScript.getMonth() + "\n");
+				writer.write(nseScript.getSymbol() + SEP_COMMA + nseScript.getDay() + SEP_COMMA + nseScript.getWeek() + SEP_COMMA
+						+ nseScript.getMonth() + SEP_COMMA +  nseScript.getSupport() + SEP_COMMA + nseScript.getResistance() + "\n");
 			}
 
 			System.out.println("Report Created is in Location : " + f.getAbsolutePath());
@@ -402,25 +456,33 @@ public class Trading_View {
 			ex.printStackTrace();
 		}
 	}
-	
-	private static long getTimeStamp()
-	{
+
+	private static long getTimeStamp() {
 		Date date = new Date();
 		Timestamp ts = new Timestamp(date.getTime());
 		System.out.println(ts);
 		long startTS = ts.getTime();
-        return startTS;
+		return startTS;
 	}
 
+	private static void printErrorList(List<String> errList)
+	{
+		System.err.println("NSE Scripts failed to process:");
+		for (Iterator iterator = errList.iterator(); iterator.hasNext();) {
+			String string = (String) iterator.next();
+			System.err.println(string);
+		}
+	}
+	
 	public static void main(String[] args) throws Exception {
-				
-		long startTS = getTimeStamp();
-        System.out.println("Start Time: " + startTS);
 
-		
+		long startTS = getTimeStamp();
+		System.out.println("Start Time: " + startTS);
+
 		try {
 
-			lastRecordSymbol = readLastRecordSymbol(outFilename);
+			if (!overwriteFlag)
+				lastRecordSymbol = readLastRecordSymbol(outFilename);
 
 //Read NSE scripts		
 			records = readFile(inputFileName);
@@ -432,14 +494,12 @@ public class Trading_View {
 				NSEScript nseScript = (NSEScript) iterator.next();
 
 				if (startCrawling) {
-					try
-					{
+					try {
 						crawlMainPage(nseScript);
-					}
-					catch(Exception ex)
-					{
+					} catch (Exception ex) {
 						System.err.println(ex.getMessage());
 						System.err.println("ERROR: Processing " + nseScript.getSymbol());
+						errorList.add(nseScript.getSymbol());
 						retryRecords.add(nseScript);
 					}
 				} else if (lastRecordSymbol != null && !lastRecordSymbol.isEmpty()
@@ -457,14 +517,26 @@ public class Trading_View {
 			writeOutputFile(outFilename, outRecords);
 
 			writeOutputFile(retryfilename, retryRecords);
+			
+			//BUY Records
+			writeOutputFile(outFilename+"buy_radar", buyRecords);
+			
+			//SELL Records
+			writeOutputFile(outFilename+"sell_radar", sellRecords);
 
+			
 			if (chrome != null)
 				chrome.quit();
+
+			//Print URL's not processed
+			printErrorList(errorList);
 			
 			long endTS = getTimeStamp();
-	        System.out.println("End Time: " + endTS);
-	        System.out.println("Time Differnce(ms): " + (endTS - startTS));
-			
+			System.out.println("End Time: " + endTS);
+			System.out.println("Total Time Taken(ms): " + (endTS - startTS));
+//			System.out.println("Average Time Taken(ms): " + ((double) (endTS - startTS) / (double) records.size()));
+			System.out.println("Average Time Taken(sec): " + ((double) (endTS - startTS) / (double) records.size() / 1000));
+
 		}
 
 	}
